Remove Passport plugin from CD model
As a result of using the plugin on the CD model, CD records had a
property added to them called `username_1`. This property was not
defined; however, it seems that Mongoose may have been enforcing a
uniqueness constraint, but not treating null values as unique. This
may explain why it is possible to create one user when signing up, but
not two.

[UPDATE] This was the issue; having the plugin attached to the CD model
cause mLab to enforce a uniqueness constraint on a property that was not
being defined; once the first user was created, there was already a null
value for that field, so attempting to create a second user (which would
also have a null value) would fail Passport's model validations.

-- GA-MEB (Matt Brendzel, 2016-10-24)